Scripts/python/gemini_images/requirements.txt
google-genai==1.36.0
pillow==11.3.0
